[Breaking] Moves JsonSerializer to datamodel.helpers package
