+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/helpers/JsonSerializer.java
@@ -1,4 +1,4 @@
-package org.wikidata.wdtk.datamodel.implementation.json;
+package org.wikidata.wdtk.datamodel.helpers;
 
 /*
  * #%L
